+++ docs/source/changelog.rst
@@ -1,5 +1,21 @@
-Change Log
-==========
+Changelog
+=========
+
+2023-11-13
+----------
+- QAOA can now create cost operator from HOBO
+
+2023-11-11
+----------
+- Add params to the history of optimization, also all the results are stored in the history
+
+2023-11-08
+----------
+- Allow to pass any keyword arguments to SciPy optimizer
+
+2023-11-06
+----------
+- Add history to SolverResult and OptimizerResult
 
 2023-10-29
 ----------
@@ -19,7 +35,7 @@ Change Log
                     "stepsize": 0.0001,
                 }
             },
-            "params_inits": params_cofing
+            "params_inits": params_config
         },
         "problem": problem_config
     }